+++ ntex/src/client/h1proto.rs
@@ -187,7 +187,6 @@ impl Stream for PlStream {
                 Err(RecvError::KeepAlive) => {
                     Err(io::Error::new(io::ErrorKind::TimedOut, "Keep-alive").into())
                 }
-                Err(RecvError::Stop) => Err(io::Error::other("Dispatcher stopped").into()),
                 Err(RecvError::WriteBackpressure) => {
                     ready!(this.io.as_ref().unwrap().poll_flush(cx, false))?;
                     continue;